Rename UpgradeMessage component to QuotaExceededMessage
