+++ projects/plugins/jetpack/extensions/plugins/ai-assistant-plugin/components/ai-image/components/ai-image-modal.tsx
@@ -9,7 +9,7 @@ import { Icon, external } from '@wordpress/icons';
  * Internal dependencies
  */
 import './ai-image-modal.scss';
-import UpgradePrompt from '../../../../../blocks/ai-assistant/components/upgrade-prompt';
+import QuotaExceededMessage from '../../../../../blocks/ai-assistant/components/quota-exceeded-message';
 import AiAssistantModal from '../../modal';
 import Carrousel, { CarrouselImages } from './carrousel';
 import UsageCounter from './usage-counter';
@@ -159,7 +159,7 @@ export default function AiImageModal( {
 							</div>
 						</div>
 						{ upgradePromptVisible && (
-							<UpgradePrompt
+							<QuotaExceededMessage
 								description={ upgradeDescription }
 								placement={ FEATURED_IMAGE_UPGRADE_PROMPT_PLACEMENT }
 								useLightNudge={ true }
